fix: adjust code
