+++ src/controller/eme-controller.ts
@@ -968,6 +968,7 @@ class EMEController extends Logger implements ComponentAPI {
           'buffer' in keyId
             ? new Uint8Array(keyId.buffer, keyId.byteOffset, keyId.byteLength)
             : new Uint8Array(keyId);
+        const oldKeyIdWithStatusChange = arrayToHex(keyIdArray);
         if (
           mediaKeySessionContext.keySystem === KeySystems.PLAYREADY &&
           keyIdArray.length === 16
@@ -982,6 +983,7 @@ class EMEController extends Logger implements ComponentAPI {
         this.log(
           `key status change "${status}" for keyStatuses keyId: ${keyIdWithStatusChange} key-session "${mediaKeySessionContext.mediaKeysSession.sessionId}"`,
         );
+        keyStatuses[oldKeyIdWithStatusChange] = status;
         keyStatuses[keyIdWithStatusChange] = status;
       },
     );
