+++ src/loader/fragment-loader.ts
@@ -1,5 +1,6 @@
 import { ErrorDetails, ErrorTypes } from '../errors';
 import { getLoaderConfigWithoutReties } from '../utils/error-helper';
+import { fakeEncryption } from '../utils/playready-workaround';
 import type { BaseSegment, Fragment, Part } from './fragment';
 import type { HlsConfig } from '../config';
 import type {
@@ -102,6 +103,13 @@ export default class FragmentLoader {
           if (context.resetIV && frag.decryptdata) {
             frag.decryptdata.iv = new Uint8Array(payload.slice(0, 16));
             payload = payload.slice(16);
+          }
+          if (
+            hls.config.requiresEncryptionInfoInAllInitSegments &&
+            frag.sn === 'initSegment'
+          ) {
+            payload = fakeEncryption(new Uint8Array(payload))
+              .buffer as ArrayBuffer;
           }
           resolve({
             frag,
